+++ example/lib/main.dart
@@ -51,6 +51,92 @@ class _MyHomePageState extends State<MyHomePage> {
                 onChanged: print,
                 popupItemDisabled: (String s) => s.startsWith('I'),
                 selectedItem: "Tunisia",
+                onBeforeChange: (a, b) {
+                  AlertDialog alert = AlertDialog(
+                    title: Text("Are you sure..."),
+                    content: Text("...you want to clear the selection"),
+                    actions: [
+                      FlatButton(
+                        child: Text("OK"),
+                        onPressed: () {
+                          Navigator.of(context).pop(true);
+                        },
+                      ),
+                      FlatButton(
+                        child: Text("NOT OK"),
+                        onPressed: () {
+                          Navigator.of(context).pop(false);
+                        },
+                      ),
+                    ],
+                  );
+
+                  return showDialog(
+                      context: context,
+                      builder: (BuildContext context) {
+                        return alert;
+                      });
+                },
+              ),
+              Divider(),
+
+              ///Menu Mode with overriden icon and dropdown buttons
+              Row(
+                children: [
+                  Expanded(
+                    flex: 2,
+                    child: DropdownSearch<String>(
+                      validator: (v) => v == null ? "required field" : null,
+                      hint: "Select a country",
+                      mode: Mode.MENU,
+                      dropdownSearchDecoration: InputDecoration(
+                        filled: true,
+                        border: UnderlineInputBorder(
+                          borderSide: BorderSide(color: Color(0xFF01689A)),
+                        ),
+                      ),
+                      showAsSuffixIcons: true,
+                      clearButtonBuilder: (_) => Padding(
+                        padding: const EdgeInsets.all(8.0),
+                        child: const Icon(
+                          Icons.clear,
+                          size: 24,
+                          color: Colors.black,
+                        ),
+                      ),
+                      dropdownButtonBuilder: (_) => Padding(
+                        padding: const EdgeInsets.all(8.0),
+                        child: const Icon(
+                          Icons.arrow_drop_down,
+                          size: 24,
+                          color: Colors.black,
+                        ),
+                      ),
+                      showSelectedItem: true,
+                      items: [
+                        "Brazil",
+                        "Italia (Disabled)",
+                        "Tunisia",
+                        'Canada'
+                      ],
+                      label: "Menu mode *",
+                      showClearButton: true,
+                      onChanged: print,
+                      popupItemDisabled: (String s) => s.startsWith('I'),
+                      selectedItem: "Tunisia",
+                    ),
+                  ),
+                  Expanded(
+                      child: TextField(
+                    decoration: InputDecoration(
+                      filled: true,
+                      labelText: "Menu mode *",
+                      border: UnderlineInputBorder(
+                        borderSide: BorderSide(color: Color(0xFF01689A)),
+                      ),
+                    ),
+                  ))
+                ],
               ),
               Divider(),
               DropdownSearch<UserModel>(
@@ -180,12 +266,14 @@ class _MyHomePageState extends State<MyHomePage> {
                             .changeSelectedItem("No");
                       },
                       child: Text("set to 'NO'")),
-                  RaisedButton(
-                      onPressed: () {
-                        _openDropDownProgKey.currentState
-                            .changeSelectedItem("Yes");
-                      },
-                      child: Text("set to 'YES'")),
+                  Material(
+                    child: RaisedButton(
+                        onPressed: () {
+                          _openDropDownProgKey.currentState
+                              .changeSelectedItem("Yes");
+                        },
+                        child: Text("set to 'YES'")),
+                  ),
                   RaisedButton(
                       onPressed: () {
                         _openDropDownProgKey.currentState