+++ lib/dropdown_search.dart
@@ -25,6 +25,8 @@ typedef Widget ErrorBuilder<T>(
     BuildContext context, String searchEntry, dynamic exception);
 typedef Widget EmptyBuilder<T>(BuildContext context, String searchEntry);
 typedef Widget LoadingBuilder<T>(BuildContext context, String searchEntry);
+typedef Widget IconButtonBuilder(BuildContext context);
+typedef Future<bool> BeforeChange<T>(T prevItem, T nextItem);
 
 enum Mode { DIALOG, BOTTOM_SHEET, MENU }
 
@@ -125,8 +127,17 @@ class DropdownSearch<T> extends StatefulWidget {
   ///custom dropdown clear button icon widget
   final Widget clearButton;
 
+  ///custom clear button widget builder
+  final IconButtonBuilder clearButtonBuilder;
+
   ///custom dropdown icon button widget
   final Widget dropDownButton;
+
+  ///custom dropdown button widget builder
+  final IconButtonBuilder dropdownButtonBuilder;
+
+  ///whether to manage the clear and dropdown icons via InputDecoration suffixIcon
+  final bool showAsSuffixIcons;
 
   ///If true, the dropdownBuilder will continue the uses of material behavior
   ///This will be useful if you want to handle a custom UI only if the item !=null
@@ -144,6 +155,14 @@ class DropdownSearch<T> extends StatefulWidget {
 
   ///called when popup is dismissed
   final VoidCallback onPopupDismissed;
+
+  /// callback executed before applying value change
+  ///delay before searching, change it to Duration(milliseconds: 0)
+  ///if you do not use online search
+  final Duration searchDelay;
+
+  /// callback executed before applying value change
+  final BeforeChange<T> onBeforeChange;
 
   DropdownSearch({
     Key key,
@@ -178,13 +197,18 @@ class DropdownSearch<T> extends StatefulWidget {
     this.autoFocusSearchBox = false,
     this.dialogMaxWidth,
     this.clearButton,
+    this.clearButtonBuilder,
     this.dropDownButton,
+    this.dropdownButtonBuilder,
+    this.showAsSuffixIcons = false,
     this.dropdownBuilderSupportsNullItem = false,
     this.popupShape,
     this.popupItemDisabled,
     this.popupBarrierColor,
     this.onPopupDismissed,
     this.searchBoxController,
+    this.searchDelay,
+    this.onBeforeChange,
   })  : assert(isFilteredOnline != null),
         assert(dropdownBuilderSupportsNullItem != null),
         assert(enabled != null),
@@ -207,6 +231,16 @@ class DropdownSearchState<T> extends State<DropdownSearch<T>> {
   void initState() {
     super.initState();
     _selectedItemNotifier.value = widget.selectedItem;
+  }
+
+  @override
+  void didUpdateWidget(DropdownSearch<T> oldWidget) {
+    final oldSelectedItem = oldWidget.selectedItem;
+    final newSelectedItem = widget.selectedItem;
+    if (oldSelectedItem != newSelectedItem) {
+      _selectedItemNotifier.value = newSelectedItem;
+    }
+    super.didUpdateWidget(oldWidget);
   }
 
   @override
@@ -239,7 +273,7 @@ class DropdownSearchState<T> extends State<DropdownSearch<T>> {
               : Text(_selectedItemAsString(data),
                   style: Theme.of(context).textTheme.subtitle1),
         ),
-        _manageTrailingIcons(data),
+        if (!widget.showAsSuffixIcons) _manageTrailingIcons(data),
       ],
     );
   }
@@ -265,7 +299,7 @@ class DropdownSearchState<T> extends State<DropdownSearch<T>> {
                     (widget.dropdownBuilder == null ||
                         widget.dropdownBuilderSupportsNullItem),
                 isFocused: isFocused,
-                decoration: _manageDropdownDecoration(state),
+                decoration: _manageDropdownDecoration(state, value),
                 child: _defaultSelectItemWidget(value),
               );
             });
@@ -274,7 +308,7 @@ class DropdownSearchState<T> extends State<DropdownSearch<T>> {
   }
 
   ///manage dropdownSearch field decoration
-  InputDecoration _manageDropdownDecoration(FormFieldState state) {
+  InputDecoration _manageDropdownDecoration(FormFieldState state, T data) {
     return (widget.dropdownSearchDecoration ??
             InputDecoration(
                 contentPadding: EdgeInsets.fromLTRB(12, 12, 0, 0),
@@ -284,6 +318,8 @@ class DropdownSearchState<T> extends State<DropdownSearch<T>> {
             enabled: widget.enabled,
             labelText: widget.label,
             hintText: widget.hint,
+            suffixIcon:
+                widget.showAsSuffixIcons ? _manageTrailingIcons(data) : null,
             errorText: state.errorText);
   }
 
@@ -300,20 +336,33 @@ class DropdownSearchState<T> extends State<DropdownSearch<T>> {
 
   ///function that manage Trailing icons(close, dropDown)
   Widget _manageTrailingIcons(T data) {
+    final clearButtonPressed = () => _handleOnChangeSelectedItem(null);
+    final dropdownButtonPressed = () => _selectSearchMode(data);
+
     return Row(
       mainAxisSize: MainAxisSize.min,
       mainAxisAlignment: MainAxisAlignment.end,
       children: <Widget>[
         if (data != null && widget.showClearButton)
-          IconButton(
-            icon: widget.clearButton ?? const Icon(Icons.clear, size: 24),
-            onPressed: () => _handleOnChangeSelectedItem(null),
-          ),
-        IconButton(
-          icon: widget.dropDownButton ??
-              const Icon(Icons.arrow_drop_down, size: 24),
-          onPressed: () => _selectSearchMode(data),
-        ),
+          widget.clearButtonBuilder != null
+              ? GestureDetector(
+                  onTap: clearButtonPressed,
+                  child: widget.clearButtonBuilder(context),
+                )
+              : IconButton(
+                  icon: widget.clearButton ?? const Icon(Icons.clear, size: 24),
+                  onPressed: clearButtonPressed,
+                ),
+        widget.dropdownButtonBuilder != null
+            ? GestureDetector(
+                onTap: dropdownButtonPressed,
+                child: widget.dropdownButtonBuilder(context),
+              )
+            : IconButton(
+                icon: widget.dropDownButton ??
+                    const Icon(Icons.arrow_drop_down, size: 24),
+                onPressed: dropdownButtonPressed,
+              ),
       ],
     );
   }
@@ -346,14 +395,12 @@ class DropdownSearchState<T> extends State<DropdownSearch<T>> {
         shape: widget.popupShape,
         context: context,
         builder: (context) {
-          return SingleChildScrollView(
-            child: AnimatedPadding(
-              duration: Duration(milliseconds: 300),
-              padding: EdgeInsets.only(
-                bottom: MediaQuery.of(context).viewInsets.bottom,
-              ),
-              child: _selectDialogInstance(data, defaultHeight: 350),
+          return AnimatedPadding(
+            duration: Duration(milliseconds: 300),
+            padding: EdgeInsets.only(
+              bottom: MediaQuery.of(context).viewInsets.bottom,
             ),
+            child: _selectDialogInstance(data, defaultHeight: 350),
           );
         });
   }
@@ -418,6 +465,7 @@ class DropdownSearchState<T> extends State<DropdownSearch<T>> {
       itemDisabled: widget.popupItemDisabled,
       searchBoxController:
           widget.searchBoxController ?? TextEditingController(),
+      searchDelay: widget.searchDelay,
     );
   }
 
@@ -433,8 +481,23 @@ class DropdownSearchState<T> extends State<DropdownSearch<T>> {
 
   ///handle on change value , if the validation is active , we validate the new selected item
   void _handleOnChangeSelectedItem(T selectedItem) {
-    _selectedItemNotifier.value = selectedItem;
-    if (widget.onChanged != null) widget.onChanged(selectedItem);
+    final changeItem = () {
+      _selectedItemNotifier.value = selectedItem;
+      if (widget.onChanged != null) widget.onChanged(selectedItem);
+    };
+
+    if (widget.onBeforeChange != null) {
+      widget
+          .onBeforeChange(_selectedItemNotifier.value, selectedItem)
+          .then((value) {
+        if (value == true) {
+          changeItem();
+        }
+      });
+    } else {
+      changeItem();
+    }
+
     _handleFocus(false);
   }
 
@@ -471,6 +534,10 @@ class DropdownSearchState<T> extends State<DropdownSearch<T>> {
   ///value PROGRAMMATICALLY, Otherwise you can use [_handleOnChangeSelectedItem]
   void changeSelectedItem(T selectedItem) =>
       _handleOnChangeSelectedItem(selectedItem);
+
+  ///Change selected Value; this function is public USED to clear selected
+  ///value PROGRAMMATICALLY, Otherwise you can use [_handleOnChangeSelectedItem]
+  void clear() => _handleOnChangeSelectedItem(null);
 
   ///get selected value programmatically
   T get getSelectedItem => _selectedItemNotifier.value;